+++ src/main/scala/tutorial/webapp/ProgramTrace.scala
@@ -11,9 +11,9 @@ case class MemoryWrite(address: Int) extends TraceEvent
 
 trait ProgramTrace {
   def read(
-      at: Int,
+      at: Long,
       last: Int,
       forward: Boolean = true
   ): Future[Seq[TraceEvent]]
-  def length(): Future[Int]
+  def length(): Future[Long]
 }
